refactor: move extractXmlFromHtmlWrapper to separate utility file
